chore(node): simplify URL and URLSearchParams global definitions
